rename *Tests to *Test
