+++ src/test/java/uk/gov/hmcts/reform/sandl/snlevents/service/HearingPartServiceTest.java
@@ -28,7 +28,7 @@ import static org.mockito.Matchers.any;
 import static org.mockito.Mockito.when;
 
 @RunWith(SpringRunner.class)
-public class HearingPartServiceTests {
+public class HearingPartServiceTest {
     @InjectMocks
     HearingPartService hearingPartService;
